Refactor how GSS authentication is organized
